feat: add custom use-countdown hook bb-321
