+++ apps/mobile/src/libs/components/deadline-countdown/libs/hooks/hooks.ts
@@ -1,0 +1,1 @@
+export { useCountdown } from "./use-countdown.hook";
